Made Significant Progress in 03_Libya_Apartment_Destroyed
There is now a cracked roof and wall, and things have fallen over.

Important Notes:
We need to add physics to the roof lights so that they swing.
We need to add physics to objects so you can knock them around or so
that they fall down at the beginning of the level.
Mist effects?